Add the ability to use InfluxDB 0.9
This no longer uses the dot notation for metric names, but instead uses
tags to specify the metadata.  The perfdata variable is used as the
metric name, the value is stored in a column names 'value', and both
host name and check name are stored as tags.  An admin also has the
ability to specify additional tags like data center location etc.