Add the ability to use InfluxDB 0.9
This no longer uses the dot notation for metric names, but instead uses
tags to specify the metadata. The service description is used as the
metric name, and the value is stored in a column named 'value'.  Both
the host name and perfdata variable name are stored as tags. An admin also
has the ability to specify additional tags like data center location
etc.

This corrects the last commit, which mistakenly mixed up what went in
the name and what went in tags when describing how it worked.